pyproject.toml
[tool.poetry]
name = "tus-datos-challange"
version = "0.0.1"
description = "Challange tus_Datos"
authors = ["hasala1996 <90461844+hasala1996@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
fastapi = "0.105.0"
sqlalchemy = "2.0.3"
pydantic = { version = "2.8.2", extras = ["email"] }
passlib = "1.7.4"
alembic = "1.9.2"
bcrypt = "3.2.0"
PyJWT = { version = "2.9.0", extras = ["crypto"] }
psycopg2-binary = "2.9.10"
requests = "2.32.3"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
uvicorn = "0.20.0"
SQLAlchemy-Utils = "0.41.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.2"
httpx = "0.27.2"
coverage = "7.6.1"
Faker = "28.4.1"
pre-commit = "3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
